+++ CHANGELOG.md
@@ -17,7 +17,7 @@ Features, Bug Fixes, Breaking Changes, Deprecated
 
 ### Features 
 
-- Support RekeyTo in the inner transaction for TEAL v6.
+- Support RekeyTo field in the inner transaction for TEAL v6.
 
 ## v3.2.0 2022-02-03